Move ToolUser implementation to ToolUser.cpp
